ai-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
spacy==3.7.2
transformers==4.36.0
sentence-transformers==2.2.2
scikit-learn==1.3.2
numpy==1.24.3
pandas==2.0.3
nltk==3.8.1
pdfplumber==0.9.0
python-docx==0.8.11
faiss-cpu==1.7.4
redis==5.0.1
python-dotenv==1.0.0
httpx==0.25.2
